refactor: replace old Path with Path2 and change domain name
- delete unused classes
