+++ src/main/java/wooteco/subway/service/PathService.java
@@ -2,7 +2,7 @@ package wooteco.subway.service;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
-import wooteco.subway.domain.path.Path2;
+import wooteco.subway.domain.path.Path;
 import wooteco.subway.domain.section.Section;
 import wooteco.subway.domain.station.Station;
 import wooteco.subway.dto.response.DtoAssembler;
@@ -25,7 +25,7 @@ public class PathService {
         Station startStation = stationRepository.findExistingStation(sourceStationId);
         Station endStation = stationRepository.findExistingStation(targetStationId);
         List<Section> sections = subwayRepository.findAllSections();
-        Path2 path = Path2.of(startStation, endStation, sections);
+        Path path = Path.of(startStation, endStation, sections);
 
         return DtoAssembler.assemble(path);
     }
